+++ CHANGELOG.md
@@ -27,6 +27,7 @@ for new features.
 - **spinner**: Change border-mode spinner design to match
   recent [progress bar design](https://designmetier-bootstrap.sncf.fr/docs/4.3/components/progress/) updates
 - **spinner** tweaked the animation timing for "growing" spinners
+- **input password** add `icon` : it is now possible to toggle text visibility from password to text by cliking on eye icon
 
 ### Deprecated